# Vectorized DeltaBitPackDecoder (#1281)