@tustvold 2022-02-12
Apologies for taking so long to get back to this, but I think this should incorporate all the feedback now