@tustvold 2022-02-13
I'm personally fairly confident in the coverage from the fuzz tests, but getting more coverage, particularly of different encoders which may interpret the spec differently, definitely couldn't hurt and would add an extra safety guarantee :+1: