@alamb 2022-02-15
I spent a few hours this morning testing this reader compared to the existing implementation on an internal corpus of parquet files using this tool: https://github.com/alamb/parquet_cmp

I am pleased to say the same values were produced:
```
107 files read with different readers compared successfully
```

I encourage everyone else with a similar corpus to give it a try and report their results

🚀