@sunchao 2022-02-15
Thank you both for the input! this information is very useful. Yes it does seem a good fit for the timestamp use case. We are also evaluating the V2 format here at Apple and trying to identify scenarios where it is clearly better than V1.